ref(core): update fbx loader to 2014.2 SDK
On newer macOS systems, the gcc stdlib is no longer provided, but the 2012.2 SDK used by Amnesia links to it. The oldest SDK that links to libc++ (Clang) is 2014.2, so I used the static library for that and had to update the mesh importer for all the updated APIs etc.